2025-10-22 Chip Handler

+++ app/src/main/java/com/example/diallog/data/model/Transcript.java
@@ -4,10 +4,10 @@ public final class Transcript {
     public final String text;
     public final long startMs;
     public final long endMs;
-    public final Float confidence;
+    public final Double confidence;
     public final String speakerLabel;
 
-    public Transcript(String text, long startMs, long endMs, Float confidence, String speakerLabel) {
+    public Transcript(String text, long startMs, long endMs, Double confidence, String speakerLabel) {
         this.text = text;
         this.startMs = startMs;
         this.endMs = endMs;
